checked valgrind

+++ account.c
@@ -10,12 +10,12 @@ int deposit(struct Account *a, float amount){
     return 1;
 }
 
-double withdraw(struct Account *a, float amount){
+float withdraw(struct Account *a, float amount){
     if(a->balance>=amount){
         a->balance-=amount;
         return amount;
     } else{
-        double ret = a->balance;
+        float ret = a->balance;
         a->balance = 0;
         return ret;
     }
